Groovy  -> Java
