--- remoting/dolphin-remoting-shared/src/main/groovy/org/opendolphin/core/ModelStoreEvent.groovy
@@ -1,52 +1,0 @@
-/*
- * Copyright 2015-2016 Canoo Engineering AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.opendolphin.core
-
-import groovy.transform.EqualsAndHashCode
-
-//CompileStatic
-@EqualsAndHashCode
-public class ModelStoreEvent<A extends Attribute, P extends PresentationModel<A>> {
-    public enum Type {
-        ADDED, REMOVED
-    }
-
-    private final Type type;
-    private final P presentationModel;
-
-    public ModelStoreEvent(Type eventType, P presentationModel) {
-        this.type = eventType;
-        this.presentationModel = presentationModel;
-    }
-
-    public Type getType() {
-        return type;
-    }
-
-    public P getPresentationModel() {
-        return presentationModel;
-    }
-
-    public String toString() {
-        return new StringBuilder()
-                .append("PresentationModel ")
-                .append(type == Type.ADDED ? "ADDED" : "REMOVED")
-                .append(" ")
-                .append(presentationModel.getId())
-                .toString();
-    }
-
-}
